Add interface permissions
This only allows defining permissions outside of the interface,
so you cannot have a mutable reference to an interface value that
is itself immutable.

We should probably change that to have interface <basePermission>,
but we have the same problem with the other reference types: map,
chan, function, slice. Not sure what to do there.