scanner: If reader is a RuneScanner, do not buffer it again
We are going to use strings.Reader. Strings.Reader implements
the io.RuneScanner interface, and that's really all we need.

So change the implementation to not hardcode a *bufio.Reader,
but use a io.RuneScanner instead, and if a reader implements
it, use it directly; otherwise wrap it in a bufio reader.

Before:
BenchmarkParser-4         200000              9482 ns/op

After:
BenchmarkParser-4         300000              5905 ns/op

A 37.7% performance improvement.