Add parser for StructPermission
Missed that on Monday, it seems. Not sure if there really
is a use case for parsing these, but we at least need them
internally.

Thinking further about this, depending on how the lookup
and semantic stuff works, the field permissions might need
to be wrapped in a FieldPermission struct (and for consistency,
the same for function parameters - which are just fields, too).