Introduce 'capabilites' package for capability checking
The capabilities package is modelled after go/types, and will
later on implement checking of capabilities. For now, all it
can do is build a map of nodes to permissions they are annotated
with.

The checker design as it is now provides the ability to configure
a list of passes that are processed after each other, each pass
processing all files in a package.