permission: Add TypeMapper and TypeMapper.NewFromType()
The NewFromType() function constructs a permission matching
a given type. The idea is that this is the maximum set of
permissions possible, and can be restricted by annotations
or context with some kind of Permission.LimitTo(Permission)
operation.

Some things are still missing, for example, how to deal with
ownership, and contextual permissions (a compat mode vs current
mode thingy).