Add permission conversion / limiting
This is similar to Capability's limit function, but
works recursively and the ownership of the target
is used. It is used for generating the initial
permission set of a capability - by first creating
a permission for the type (potentially with cycles)
and then applying any annotated restrictions to
that.

Errors are handled in a panic()/recover() fashion
internally, as it's hard to read with ifs and all
the recursion (like parser and scanner).

For interfaces, it is a bit unclear if we should propagate
the outer annotation to the inner methods. If we mark an
interface as r, should we still be able to call methods
that are themselves m (that is, can return different
results in different invocations). Maybe, but let's
play safe for now.

Converting pointer types is tricky. If the new outer value
loses linearity, we need to remove linearity from the pointed
to value - it becomes shared at that point.

For linear reads, we just drop the linearity - reads are
considered to be non-destructive anyway (we only need them
to have non-writable linear pointer values). For writes, we
also drop the write permission itself, as writes really
are destructive operations.