Share a single state map between assignable checkers
This might fix some problems with recursion, where one
kind of assignment invoked another and that might have
invoked the original one again. It's unclear if that
can actually happen, but better safe than sorry.

Eventually, the three assignability checkers should
be merged into one and that should just decide based
on the assignableMode, so this can be seen as a first
step to refactor this.