Add support for merging permissions
Given two permissions, allow calculating an interesection and a
union of them. Intersection is used for joining different code
branches. If in one branch v has permission A, and in another it
has B, then after that branches it will have intersect(A, B).

Union exists mostly for function parameters, as they are contra-
variant, basically.