Initial interpreter state store implementation
Not entirely clear yet how we deal with package "state" (really
must remain the same before / after a function) and references
to other packages.

The effective vs maximum effectively allows us to punch holes into
data structures for borrowed parts by removing any maximum permissions
for the part of the data structure that was borrowed. When the borrowed
reference exceeded its life time, the old permission can be restored.
It's not on the roadmap to actually add this, but it seems worthwhile
to have a more generic approach than just a borrowed flag or something.