Add dependencies between expressions
When evaluating something like f((a)) where a is mutable and
f takes an owned mutable argument, f should consume a. If we
just evaluate the tree however we could have an intermediate
paren node and would not know that a needs to be consumed.

By introducing dependencies between expressions we can specify
the the expression (a) depends on a, and we can then consume
the dependencies of (a) when evaluating f((a)).