Introduce wildcards as neutral element for merge
Wildcard permissions essentially are the neutral element
for merging (and converting), but do not have any assignability
rules, as it would be unclear what the behaviour should be when
such an incomplete permission specification exists.

They should not appear in the interpreter, but only in annotations:
When the annotation is merged with a complete permission matching
the type, the wildcard essentially means to just take the type's
default permission.

Given that wildcards do not support assignability it is conceivable
that the interpreter might use a variant thereof for marking borrowed
variables, though, as wildcards act as some kind of bottom value
for assignability.

Fixes #7