Do not change method base permission for interfaces
Functions are special: Their mutability flag means that the function
can return different values in different invocations (essentially
meaning that it has captured a mutable value). We cannot restrict
the functions base permission, as we would also have to freeze its
behavior. While we could theoretically freeze a function by using
memoization, that is both inefficient and would be a change to
language semantics: We would have to implement a preprocessor
for it.

Also add tests for other cases of linear stuff containing mutable
stuff.