interpreter/binary: Release earlier, and add short-circuit
Evaluating both sides first and then release their dependencies
means that it was not possible to evaluate `a && a`, for example,
because the LHS would take `a` away and it would not be usable in
the RHS. We instead need to release deps immediately after evaluating
a side.

Also add support for short-circuit logic. In A && B and A || B, only
A could be evaluated, hence the result is not the result of evaluating
B after A, but could be both. This is represented by joining the two
possible branches using an intersection.