permission: Require Read on source when copying
This makes somewhat sense. We still cannot require Write
on the target, though, as we need to be able to perform
the checks for initialization of read-only variables.

This should be extended to move, but not to refcopy, because
refcopy sucks.

Fixes an unsoundness where n * r could be copied to r * r.