interpreter: Add support for composite literals
This is solved suboptimally with regards to unowned
composite literals, as it uses the []Borrowed infrastructure
correctly, but []Borrowed kind of became a way to specify
owners of the object. That probably needs to be split; this
would also open up the way for functions to return unowned
values (which depend on the (unowned) input arguments).

There is another issue in that types and values share the
same namespace with permissions of the same kind, so a type
T and a value v : T could have the same permission. It probably
makes sense for these cases to require a

	type TypePermission struct { Instance Permission}

so we can differentiate these categories; although it is not
a problem in practice, as Go's type checker would not allow
us to use a type in a value context and vice versa.