Add NilPermission
Values of the `nil` literal are represented by the an 'untyped
nil' basic type. Basic types where converted to base permissions
in the typemapper, causing them to not have pointer types.

It would have been nice to just directly assign the correct
pointer permission when accessing the nil identifier, but this
does not seem easily possible, so this seems to be the best
way forward.