(Normal) function calls should not have an owner
A function call's return value should not have an owner, it's
completely unclear which argument it was derived from, and it
really makes no sense to make the function the owner.

We need to have the function as a owner for defered and go
statements, though, as they have different binding requirements
- like defer obj.f() should consume obj, but defer func() { obj }()
should not.