interpreter: Differentiate between owner and other dependencies
Any expression can have an owner (for example, a.b has the owner
a, and a.b() might have the owner a), and dependencies.

A function call's return value should not have an owner, it's
completely unclear which argument it was derived from, and it
really makes no sense to make the function the owner.

We need to have the function as a owner for defered and go
statements, though, as they have different binding requirements
- like defer obj.f() should consume obj, but defer func() { obj }()
should not.

This also fixes a problem we'd have hit eventually with
"freezing" values. When a linear mutable value is moved
to a non-linear value, it is frozen. This was implemented
by freezing the dependencies, but that's wrong.