permission: Merge assignableCopy as well
Annotate permissions for reference types using a function called
copyAsReference(), so copying them involves actually referencing
them; and also use that for the target of a pointer.

This is the final step in unifying all assignable checks into
one implementation, severily reducing confusion and potential
incompatible changes in the recursive implementations of them.