Rework non-entering for switch/select statement
In the switch statement, we added the store before evaluating the
tag to the output, which is false. Also, if the body was empty, we
produced two exits, because visitStmtList for an empty block returns
one exit, with the input store.

Solve the problem by removing the generation of non-entering exits
from select and switch, and add the non-entering exit from visitStmtList,
if isASwitch is true.