int: Pass owner, deps to defineOrAssign()
While this is basically NoOwner, nil everywhere right now, this
actually makes no sense if you look at the function in isolation.

It's much easier to explain the function if it makes sense on its
own, though, so let's fix this.

Also note that assigning to _ releases owner and dependencies -
it essentially is a no-op.